Move quantiles/pareto computation to the Video model
